# Source/subscriptions-swagger.yaml
# YAML for swagger definition of the subscription service
swagger: '2.0'

# This is your document metadata
info:
  version: "0.1.0"
  title: Subscription Service

# Describe your paths here
paths:
  # Subscriptions
  /subscription/{subscriptionId}:
    # This is a HTTP operation
    get:
      # Describe this verb here. Note: you can use markdown
      description: |
        Gets `Subscription` object by id.
      # This is array of GET operation parameters:
      operationId: getSubscriptionById
      produces:
        - application/json
        - application/xml      
      parameters:
        - in: path
          name: subscriptionId
          description: ID of subscription that needs to be fetched
          required: true
          type: integer
          format: int64
      # Expected responses for this operation:
      responses:
        "200":
          description: successful operation
          schema:
            $ref: "#/definitions/Subscription"
        "400":
          description: Invalid subscription id value
  # All Subscriptions
  /subscriptions:
    # This is a HTTP operation
    get:
      description: |
        Gets all `Subscription` objects.
        Optional query param of **size** determines
        size of returned array
      # This is array of GET operation parameters:
      operationId: getAllSubscriptions
      produces:
        - application/json
        - application/xml      
      # Expected responses for this operation:
      responses:
        "200":
          description: successful operation
          schema:
            type: array
            items:
              $ref: "#/definitions/Subscription"
        "400":
          description: Invalid subscription id value          
definitions:
  Subscription:
    type: object
    properties:
      id:
        type: integer
        format: int64
      subscriberId:
        type: integer
        format: int64
      subscriptionType:
        type: string
      startDate:
        type: string
        format: date
      endDate:
        type: string
        format: date
      valid:
        type: boolean